Begin to separate scheme definition from assembly
