+++ include/samurai/petsc.hpp
@@ -4,11 +4,11 @@
 
 #pragma once
 #include "numeric/error.hpp"
-#include "petsc/fv/diffusion_FV.hpp"
-#include "petsc/fv/diffusion_FV_old.hpp"
-#include "petsc/fv/divergence_FV.hpp"
-#include "petsc/fv/gradient_FV.hpp"
-#include "petsc/fv/identity_FV.hpp"
-#include "petsc/fv/zero_operator_FV.hpp"
-#include "petsc/scheme_operators.hpp"
+// #include "petsc/scheme_operators.hpp"
 #include "petsc/solver.hpp"
+#include "schemes/fv/diffusion_FV.hpp"
+#include "schemes/fv/diffusion_FV_old.hpp"
+#include "schemes/fv/divergence_FV.hpp"
+#include "schemes/fv/gradient_FV.hpp"
+#include "schemes/fv/identity_FV.hpp"
+#include "schemes/fv/zero_operator_FV.hpp"
