+++ include/samurai/petsc/fv/flux_based_scheme_assembly.hpp
@@ -1,169 +1,54 @@
 #pragma once
-#include "../interface.hpp"
-#include "fv/FV_scheme.hpp"
+#include "../../interface.hpp"
+#include "FV_scheme_assembly.hpp"
 
 namespace samurai
 {
     namespace petsc
     {
-        /**
-         * Defines how to compute a normal flux: e.g., Grad(u).n
-         * - direction: e.g., right = {0, 1}
-         * - stencil: e.g., current cell and right neighbour = {{0, 0}, {0, 1}}
-         * - function returning the coefficients for the flux computation w.r.t. the stencil:
-         *            auto get_flux_coeffs(double h)
-         *            {
-         *                // Grad(u).n = (u_1 - u_0)/h
-         *                std::array<double, 2> flux_coeffs;
-         *                flux_coeffs[0] = -1/h; // current cell
-         *                flux_coeffs[1] =  1/h; // right neighbour
-         *                return flux_coeffs;
-         *            }
-         */
-        template <class Field, std::size_t stencil_size>
-        struct NormalFluxComputation
+        template <class Scheme>
+        class FluxBasedSchemeAssembly : public FVSchemeAssembly<Scheme>
         {
-            static constexpr std::size_t dim        = Field::dim;
-            static constexpr std::size_t field_size = Field::size;
-            using field_value_type                  = typename Field::value_type;                      // double
-            using flux_matrix_t = typename detail::LocalMatrix<field_value_type, 1, field_size>::Type; // 'double' if field_size = 1,
-                                                                                                       // 'xtensor' representing a matrix
-                                                                                                       // otherwise
-            using flux_coeffs_t = std::array<flux_matrix_t, stencil_size>;
+            // template <class Scheme1, class Scheme2>
+            // friend class FluxBasedScheme_Sum_CellBasedScheme;
 
-            DirectionVector<dim> direction;
-            Stencil<stencil_size, dim> stencil;
-            std::function<flux_coeffs_t(double)> get_flux_coeffs;
-        };
+            // template <class Scheme>
+            // friend class Scalar_x_FluxBasedScheme;
 
-        template <class Field, class Vector>
-        auto normal_grad_order2(Vector& direction)
-        {
-            static constexpr std::size_t dim        = Field::dim;
-            static constexpr std::size_t field_size = Field::size;
-            using flux_computation_t                = NormalFluxComputation<Field, 2>;
-            using flux_matrix_t                     = typename flux_computation_t::flux_matrix_t;
-
-            flux_computation_t normal_grad;
-            normal_grad.direction       = direction;
-            normal_grad.stencil         = in_out_stencil<dim>(direction);
-            normal_grad.get_flux_coeffs = [](double h)
-            {
-                std::array<flux_matrix_t, 2> coeffs;
-                if constexpr (field_size == 1)
-                {
-                    coeffs[0] = -1 / h;
-                    coeffs[1] = 1 / h;
-                }
-                else
-                {
-                    coeffs[0].fill(-1 / h);
-                    coeffs[1].fill(1 / h);
-                }
-                return coeffs;
-            };
-            return normal_grad;
-        }
-
-        template <class Field>
-        auto normal_grad_order2()
-        {
-            static constexpr std::size_t dim = Field::dim;
-            using flux_computation_t         = NormalFluxComputation<Field, 2>;
-
-            auto directions = positive_cartesian_directions<dim>();
-            std::array<flux_computation_t, dim> normal_fluxes;
-            for (std::size_t d = 0; d < dim; ++d)
-            {
-                normal_fluxes[d] = normal_grad_order2<Field>(xt::view(directions, d));
-            }
-            return normal_fluxes;
-        }
-
-        template <class Field, std::size_t output_field_size, std::size_t stencil_size>
-        struct FluxBasedCoefficients
-        {
-            static constexpr std::size_t dim        = Field::dim;
-            static constexpr std::size_t field_size = Field::size;
-
-            using flux_computation_t = NormalFluxComputation<Field, stencil_size>;
-            using field_value_type   = typename Field::value_type; // double
-            using coeff_matrix_t     = typename detail::LocalMatrix<field_value_type, output_field_size, field_size>::Type;
-            using cell_coeffs_t      = std::array<coeff_matrix_t, stencil_size>;
-            using flux_coeffs_t      = typename flux_computation_t::flux_coeffs_t; // std::array<flux_matrix_t, stencil_size>;
-
-            flux_computation_t flux;
-            std::function<cell_coeffs_t(flux_coeffs_t&, double, double)> get_cell1_coeffs;
-            std::function<cell_coeffs_t(flux_coeffs_t&, double, double)> get_cell2_coeffs;
-        };
-
-        /**
-         * Useful sizes to define the sparsity pattern of the matrix and perform the preallocation.
-         */
-        template <PetscInt output_field_size_, PetscInt stencil_size_>
-        struct FluxBasedAssemblyConfig
-        {
-            static constexpr PetscInt output_field_size = output_field_size_;
-            static constexpr PetscInt stencil_size      = stencil_size_;
-        };
-
-        template <class cfg, class bdry_cfg, class Field>
-        class FluxBasedScheme : public FVScheme<Field, cfg::output_field_size, bdry_cfg>
-        {
-            template <class Scheme1, class Scheme2>
-            friend class FluxBasedScheme_Sum_CellBasedScheme;
-
-            template <class Scheme>
-            friend class Scalar_x_FluxBasedScheme;
-
-            template <std::size_t rows, std::size_t cols, class... Operators>
-            friend class MonolithicBlockAssembly;
+            // template <std::size_t rows, std::size_t cols, class... Operators>
+            // friend class MonolithicBlockAssembly;
 
           protected:
 
-            using base_class = FVScheme<Field, cfg::output_field_size, bdry_cfg>;
+            using base_class = FVSchemeAssembly<Scheme>;
             using base_class::cell_coeff;
             using base_class::col_index;
             using base_class::dim;
             using base_class::field_size;
             using base_class::mesh;
             using base_class::row_index;
+            using base_class::scheme;
             using base_class::set_current_insert_mode;
             using base_class::set_is_row_not_empty;
-            using dirichlet_t = typename base_class::dirichlet_t;
-            using neumann_t   = typename base_class::neumann_t;
 
           public:
 
-            using cfg_t                                    = cfg;
-            using bdry_cfg_t                               = bdry_cfg;
-            using field_t                                  = Field;
-            static constexpr std::size_t output_field_size = cfg::output_field_size;
-            static constexpr std::size_t stencil_size      = cfg::stencil_size;
+            using scheme_t                                 = Scheme;
+            using cfg_t                                    = typename Scheme::cfg_t;
+            using bdry_cfg_t                               = typename Scheme::bdry_cfg;
+            using field_t                                  = typename Scheme::field_t;
+            static constexpr std::size_t output_field_size = cfg_t::output_field_size;
+            static constexpr std::size_t stencil_size      = cfg_t::stencil_size;
 
-            using coefficients_t = FluxBasedCoefficients<Field, output_field_size, stencil_size>;
-
-          protected:
-
-            std::array<coefficients_t, dim> m_scheme_coefficients;
-
-          public:
-
-            FluxBasedScheme(Field& unknown, const std::array<coefficients_t, dim>& scheme_coefficients)
-                : base_class(unknown)
-                , m_scheme_coefficients(scheme_coefficients)
+            FluxBasedSchemeAssembly(const Scheme& scheme)
+                : base_class(scheme)
             {
                 set_current_insert_mode(ADD_VALUES);
             }
 
-            auto& scheme_coefficients() const
+            auto scheme_coefficients() const
             {
-                return m_scheme_coefficients;
-            }
-
-            auto& scheme_coefficients()
-            {
-                return m_scheme_coefficients;
+                return scheme().coefficients();
             }
 
           public:
@@ -176,7 +61,7 @@ namespace samurai
             {
                 for (std::size_t d = 0; d < dim; ++d)
                 {
-                    auto scheme_coeffs_dir = m_scheme_coefficients[d];
+                    auto scheme_coeffs_dir = scheme_coefficients()[d];
                     for_each_interior_interface(
                         mesh(),
                         scheme_coeffs_dir.flux.direction,
@@ -247,7 +132,7 @@ namespace samurai
 
                 for (std::size_t d = 0; d < dim; ++d)
                 {
-                    auto scheme_coeffs_dir = m_scheme_coefficients[d];
+                    auto scheme_coeffs_dir = scheme_coefficients()[d];
                     for_each_interior_interface(
                         mesh(),
                         scheme_coeffs_dir.flux.direction,
